refactor: simplify changing the test settings (#508)
Replace the current way of handling test settings with explicit tests.
The settings can be enabled by explicitely running the
[Testably.Abstractions.TestSettings](https://github.com/Testably/Testably.Abstractions/tree/main/Tests/Settings/Testably.Abstractions.TestSettings)
tests:
- `LongRunningTestsAlsoInDebugMode` (`Include`/`Exclude`)
In order to increase the test execution speed during development,
long-running tests are disabled per default in DEBUG mode. With this
setting, the corresponding tests can be included.
- `RealFileSystemTestsInDebugMode` (`Enable`/`Disable`)
If enabled, the classes for executing tests against the real file system
also run in DEBUG mode.

Note: This settings are stored locally in test.settings.json which is
excluded in
[.gitignore](https://github.com/Testably/Testably.Abstractions/blob/main/.gitignore)
so that it only affects the individual developer!

*Previously:*
> On the build system, unit tests are executed both against the
MockFileSystem and the RealFileSystem. This ensures that the tests
verify correct assumptions. In order to simplify and speedup the
development process, per default some tests are disabled in DEBUG mode.
These can be enabled by editing the sample
[Directory.Build.props](https://github.com/Testably/Testably.Abstractions/blob/main/Directory.Build.props)
file in the root directory and set the corresponding constants:
> 
> -
[ENABLE_REALFILESYSTEMTESTS_IN_DEBUG](https://github.com/Testably/Testably.Abstractions/blob/main/Tests/Helpers/Testably.Abstractions.Tests.SourceGenerator/ClassGenerators/FileSystemClassGenerator.cs#L59)
> If set, the classes for executing tests against the real file system
are created also in DEBUG mode.
> -
[INCLUDE_LONGRUNNING_TESTS_ALSO_IN_DEBUG_MODE](https://github.com/Testably/Testably.Abstractions/blob/main/Tests/Helpers/Testably.Abstractions.TestHelpers/Test.cs#L51)
> In order to increase the test execution speed during development, long
running tests are disabled in DEBUG mode. When this constant is set, the
corresponding tests are no longer skipped.
> -
[SKIP_TESTS_ON_REAL_FILESYSTEM](https://github.com/Testably/Testably.Abstractions/blob/main/Tests/Helpers/Testably.Abstractions.TestHelpers/Test.cs#L61)
> In order to avoid unnecessary side-effects on development machines,
tests against the real file system are not executed in DEBUG mode. When
this constant is set, this restriction no longer applies.
> **Example:**
> 
> ```xml
> <Project>
>     <PropertyGroup>
>
<DefineConstants>$(DefineConstants);ENABLE_REALFILESYSTEMTESTS_IN_DEBUG</DefineConstants>
>     </PropertyGroup>
> </Project>
> ```
> 
> *Note: This file is excluded in
[`.gitignore`](https://github.com/Testably/Testably.Abstractions/blob/main/.gitignore)
and remains local!*