fix: do not adjust LastAccessTime in `FileStream.EndRead` (#522)
Fix incorrect adjustment of the LastAccessTime in `FileStream.EndRead`
and activate the skipped brittle tests.
Refactor assertions for `DateTime` between two times by adding an
extension method `BeBetween` which checks for the time being between
start and end and applying the system clock tolerance on both values.

*Note: The test keeps failing arbitrarily on MacOS, as the
LastAccessTime is sometimes adjusted, so the test remains skipped on
MacOS*

---------

Co-authored-by: Valentin <valentin.breuss@baur.at>